feature unification workaround
