+++ Cargo.toml
@@ -2,12 +2,7 @@
 resolver = "2"
 members = [
     "bin/*",
-    "crates/core",
-    "crates/core-ethereum",
-    "crates/core-optimism",
-    "crates/preflight",
-    "crates/preflight-ethereum",
-    "crates/preflight-optimism",
+    "crates/*",
     "guests",
 ]
 
@@ -64,6 +59,7 @@ reth-revm = { git = "https://github.com/risc0/reth", branch = "p1.1.0_zstd", def
 reth-storage-errors = { git = "https://github.com/risc0/reth", branch = "p1.1.0_zstd", default-features = false, features = ["std"] }
 
 # Zeth
+zeth = { path = "crates/zeth" }
 zeth-core = { path = "crates/core" }
 zeth-core-ethereum = { path = "crates/core-ethereum" }
 zeth-core-optimism = { path = "crates/core-optimism" }
