+++ bin/zeth-ethereum/data/ethereum/1.json.gz
