+++ bin/zeth-ethereum/data/ethereum/15050000.json.gz
