+++ bin/zeth-ethereum/data/ethereum/19426587.json.gz
