+++ bin/zeth-optimism/src/main.rs
@@ -1,0 +1,29 @@
+// Copyright 2024 RISC Zero, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use tokio::task::spawn_blocking;
+use zeth_guests::{ZETH_GUESTS_RETH_OPTIMISM_ELF, ZETH_GUESTS_RETH_OPTIMISM_ID};
+use zeth_preflight_optimism::OpRethBlockBuilder;
+
+#[tokio::main]
+async fn main() -> anyhow::Result<()> {
+    spawn_blocking(move || {
+        zeth::run::<OpRethBlockBuilder, _, _, _, _, _>(
+            ZETH_GUESTS_RETH_OPTIMISM_ELF,
+            ZETH_GUESTS_RETH_OPTIMISM_ID,
+            "ethereum",
+        )
+    })
+    .await?
+}
