+++ crates/zeth/Cargo.toml
@@ -7,16 +7,10 @@ edition = "2021"
 workspace = true
 features = ["std", "client"]
 
-[dependencies.zeth-core]
-workspace = true
-
-[dependencies.zeth-preflight-ethereum]
-workspace = true
-
-[dependencies.zeth-preflight-optimism]
-workspace = true
-
 [dependencies.zeth-guests]
+workspace = true
+
+[dependencies.zeth-core]
 workspace = true
 
 [dependencies.zeth-preflight]
